refactor frontend folder structure
