+++ frontend/src/components/common/SideBar.tsx
@@ -16,7 +16,7 @@ import { useNavigate } from "react-router-dom";
 
 import * as Routes from "../../constants/routes";
 import { ReactComponent as Logo } from "../../assets/marillacPlaceLogo.svg";
-import authAPIClient from "../../APIClients/AuthAPIClient";
+import authAPIClient from "../../gql/AuthAPIClient";
 import AuthContext from "../../contexts/AuthContext";
 
 const mockAuthenticatedUser = {
