+++ frontend/src/gql/mutations/NotificationMutations.ts
