+++ frontend/src/gql/mutations/ResidentsMutations.ts
