+++ frontend/src/gql/queries/NotificationQueries.ts
