+++ frontend/src/gql/queries/ResidentsQueries.ts
