+++ frontend/src/types/Types/ResidentsType.ts
